added jackson and logging. Refactor domain model

+++ src/edu/core/java/auction/repository/BidRepository.java
@@ -6,8 +6,7 @@ import edu.core.java.auction.vo.BidValueObject;
  * Created by Max on 06.03.2017.
  */
 public class BidRepository extends Repository<BidValueObject> {
-    protected static Long maxId = (long)0;
-
+    protected static Long maxId = 0L;
 
     @Override
     public Long getMaxId(){
