+++ src/edu/core/java/auction/repository/BuyerRepository.java
@@ -6,7 +6,7 @@ import edu.core.java.auction.vo.BuyerValueObject;
  * Created by Max on 06.03.2017.
  */
 public class BuyerRepository extends Repository<BuyerValueObject> {
-    protected static Long maxId = (long)0;
+    protected static Long maxId = 0L;
 
     @Override
     public Long getMaxId(){
