+++ src/edu/core/java/auction/repository/LotRepository.java
@@ -6,7 +6,7 @@ import edu.core.java.auction.vo.LotValueObject;
  * Created by Max on 06.03.2017.
  */
 public class LotRepository extends Repository<LotValueObject> {
-    protected static Long maxId = (long)0;
+    protected static Long maxId = 0L;
 
     @Override
     public Long getMaxId(){
