+++ src/edu/core/java/auction/repository/SellerRepository.java
@@ -6,7 +6,7 @@ import edu.core.java.auction.vo.SellerValueObject;
  * Created by Max on 06.03.2017.
  */
 public class SellerRepository extends Repository<SellerValueObject> {
-    protected static Long maxId = (long)0;
+    protected static Long maxId = 0L;
 
     @Override
     public Long getMaxId(){
